requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
Django==5.1.1
django-filter==24.3
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
PyJWT==2.9.0
sqlparse==0.5.1
